refactoring - create sub directory
