+++ smtm/controller/controller.py
@@ -1,15 +1,15 @@
 """시스템 운영 인터페이스로서 Operator를 사용해서 자동 거래 시스템을 컨트롤하는 Controller 클래스"""
 
 import signal
-from .config import Config
-from .log_manager import LogManager
-from .analyzer import Analyzer
-from .upbit_trader import UpbitTrader
-from .upbit_data_provider import UpbitDataProvider
-from .bithumb_trader import BithumbTrader
-from .bithumb_data_provider import BithumbDataProvider
-from .strategy_factory import StrategyFactory
-from .operator import Operator
+from ..config import Config
+from ..log_manager import LogManager
+from ..analyzer import Analyzer
+from ..trader.upbit_trader import UpbitTrader
+from ..data.upbit_data_provider import UpbitDataProvider
+from ..trader.bithumb_trader import BithumbTrader
+from ..data.bithumb_data_provider import BithumbDataProvider
+from ..strategy.strategy_factory import StrategyFactory
+from ..operator import Operator
 
 
 class Controller:
