+++ smtm/data/bithumb_data_provider.py
@@ -3,7 +3,7 @@
 from datetime import datetime, timezone, timedelta
 import requests
 from .data_provider import DataProvider
-from .log_manager import LogManager
+from ..log_manager import LogManager
 
 
 class BithumbDataProvider(DataProvider):
