+++ smtm/strategy/strategy.py
