+++ smtm/strategy/strategy_factory.py
