+++ smtm/strategy/strategy_rsi.py
@@ -4,8 +4,8 @@ import math
 from datetime import datetime
 import numpy as np
 from .strategy import Strategy
-from .log_manager import LogManager
-from .date_converter import DateConverter
+from ..log_manager import LogManager
+from ..date_converter import DateConverter
 
 
 class StrategyRsi(Strategy):
