+++ smtm/trader/trader.py
