+++ smtm/trader/upbit_trader.py
@@ -10,9 +10,9 @@ from urllib.parse import urlencode
 import requests
 import jwt  # PyJWT
 from dotenv import load_dotenv
-from .log_manager import LogManager
+from ..log_manager import LogManager
 from .trader import Trader
-from .worker import Worker
+from ..worker import Worker
 
 load_dotenv()
